Apply silicon calibration with MIPS
Used 200 GeV mu- and looked at total energy deposited in silicon for CEE
layer 24 and CHE layer 4. The peak is used as a reference for the MIP
signal in silicon. After that apply calibration per layer in Event
Action. Adding a simple macro card to inspect the Landau distribution.
In default macro card change to negative polarity particles.