Update analysis to new selecting cuts
Discard events interacting in the first two layers of CEE, or not
interacing both in CEE and CHE or with interacting layer not found. Use
CEEInteracted variable to select CEE and CHE events. Perform gaussian fit to
extract response as the pick value of distributions (as written in TB
paper) by fitting around maximum of histograms.